Make base_note_offset a 16-bit integer rather than an 8-bit integer.

+++ src/i_oplmusic.c
@@ -64,7 +64,7 @@ typedef struct
     byte feedback;
     genmidi_op_t carrier;
     byte unused;
-    byte base_note_offset;
+    short base_note_offset;
 } PACKEDATTR genmidi_voice_t;
 
 typedef struct
